# Improved search in Filter