--- a/docs/index.html
+++ b/docs/index.html
@@ -392,7 +392,7 @@
 
 			displayFilteredPanel();
 
-			updateLink(v);
+			updateLink( v );
 
 		}