4 improvements:

1) both RegExp and full text search are supported, use /syntax/ for regexp.

```
        examples:
        - /./ => finds everything
        - /^pro/ => finds PropertyBinding + PropertyMixer, but NOT: LightProbe
        - /mesh.*material/
        - /audio|video/ => finds all audio and video stuff
        - /3$/ => finds everything that ends with a 3: CatmullRomCurve3, ..., Box3, Line3, ...
```

2) much faster processing thanks to code optimization, like:
- optimized data structures for the search => reduction of text processing
- reduction of querySelectors, those values are cached
- reduction of useless innerHTML = ... (which are costly even when assigning the same value)
- full text search is faster because it's not using RegExp anymore
- the HTML itself is generated much faster too

According to my own quick benchmarks:
- docs search is now about 6x faster, going from 37ms to 6ms
- examples search is now about 25x faster!, going from 72ms to 3ms

72ms might not seem like a lot, but keep in mind that the search is run for every single character that you enter, so if you type something that is 10 letters long, that's already 720ms instead of 30ms, and you're going to feel that lag.

3) docs/index.html and examples/index.html now share a files/search.js with common functions, including the search function.

Note: I think it would be even better to have only 1 index.html file that handles both docs and examples, because the switch between docs and examples always takes some time, and with 1 file, it would be instantaneous.

4) no more "unexpected results" due to the RegExp, and no hacks were necessary.

I haven't tested everything, so there might be a bug or two. It seems fine but please check.